LayoutTests/fast/xmlhttprequest/xmlhttprequest-sync-disabled-expected.txt

CONSOLE MESSAGE: Synchronous XMLHttpRequests are disabled for this page.
This tests that synchronous XMLHttpRequests fail when they are disabled for documents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS xhr.open("GET", "http://mydomain/", false) threw exception Error: InvalidAccessError: DOM Exception 15.
PASS successfullyParsed is true

TEST COMPLETE
